ankys-kitchen/src/app/components/ingredients/ingredients.component.spec.ts
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ComponentFixture, fakeAsync, TestBed } from '@angular/core/testing';
import { IngredientsComponent } from './ingredients.component';
import { RecipeService } from 'src/app/services/recipe.service';
import { of } from 'rxjs';

describe('IngredientsComponent', () => {
  let component: IngredientsComponent;
  let fixture: ComponentFixture<IngredientsComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [
        HttpClientTestingModule
      ],
      declarations: [ IngredientsComponent ]
    })
    .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(IngredientsComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should return the first 95 characters of the given string if there are equel/more than 95 characters', () => {
    const char100 = 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa';
    const char95 = 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa';
    const result = component.shortenDescription(char100);
    expect(result).toBe(char95);
  });

  it('should return the complete string if there are less than 95 characters', () => {
    const result = component.shortenDescription('aaaaaaaaa');
    expect(result).toBe('aaaaaaaaa');
  });

  it('should call getAllIngredients and get response', fakeAsync(() => {
    const mockData = {meals: [
      {
        "idIngredient": "1",
        "strIngredient": "Chicken",
        "strDescription": "The chicken is a type of domesticated fowl, a subspecies of the red junglefowl (Gallus gallus). It is one of the most common and widespread domestic animals, with a total population of more than 19 billion as of 2011. There are more chickens in the world than any other bird or domesticated fowl. Humans keep chickens primarily as a source of food (consuming both their meat and eggs) and, less commonly, as pets. Originally raised for cockfighting or for special ceremonies, chickens were not kept for food until the Hellenistic period (4th\u20132nd centuries BC).\r\n\r\nGenetic studies have pointed to multiple maternal origins in South Asia, Southeast Asia, and East Asia, but with the clade found in the Americas, Europe, the Middle East and Africa originating in the Indian subcontinent. From ancient India, the domesticated chicken spread to Lydia in western Asia Minor, and to Greece by the 5th century BC. Fowl had been known in Egypt since the mid-15th century BC, with the \"bird that gives birth every day\" having come to Egypt from the land between Syria and Shinar, Babylonia, according to the annals of Thutmose III.",
        "strType": null
      }
    ]}
    const service = fixture.debugElement.injector.get(RecipeService);
    spyOn(service, 'getAllIngredients').and.callFake(() => {
      return of(mockData);
    });
    component.getAllIngredients();
    expect(component.ingredients).toEqual(mockData['meals']);
  }));

});
